+++ src/ui-win/DlgSettings.cpp
@@ -30,6 +30,7 @@
 #include "DlgSettings.h"
 #include "resource/resource.h"
 #include "common/Config.h"
+#include "gup/ObjCommon.h"
 
 namespace ui {
 namespace win {
@@ -42,15 +43,26 @@ namespace win {
         //-------------------------
         mLblSimPath.setupChild(mDialog, IDC_SIM_PATH);
         mBtnOk.setupChild(mDialog, BTN_OK);
+        mChkDatarefsId.setupChild(mDialog, IDC_CHK_DATAREFS_ID);
+        mChkCommandsId.setupChild(mDialog, IDC_CHK_COMMANDS_ID);
         mBtnSelectSimPath.setupChild(mDialog, BTN_SELECT_SIM_PATH);
         //-------------------------
         mBtnOk.onClick = [&](auto) { mDialog.destroy(1); };
         mBtnSelectSimPath.onClick = [&](auto) { selectSimPath(); };
+        mChkDatarefsId.onStateChanged = [](auto,const bool state) { ObjCommon::instance()->pSettings.setUseDatarefsId(state); };
+        mChkCommandsId.onStateChanged = [](auto, const bool state) { ObjCommon::instance()->pSettings.setUseCommandsId(state); };
         //-------------------------
         mDialog.onInit = [&](auto win) {
             win->centerByParent();
             auto config = Config::instance();
+            auto & settings = ObjCommon::instance()->pSettings;
             mLblSimPath.setText(config->simDir().GetString().data());
+            mChkDatarefsId.setChecked(settings.isUseDatarefsId());
+            mChkCommandsId.setChecked(settings.isUseCommandsId());
+
+            mLblSimPath.setToolTip(_T("The path where X-Plane is installed."));
+            mChkDatarefsId.setToolTip(_T("Enabling ID usage for the project's DataRefs. The state is saved with the 3Ds Max's scene."));
+            mChkCommandsId.setToolTip(_T("Enabling ID usage for the project's Commands. The state is saved with the 3Ds Max's scene."));
         };
         //-------------------------
         mDialog.create(GetCOREInterface()->GetMAXHWnd(), DLG_SETTINGS);