+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Change-log
 
 Next
 -------------------------
+- Added: Settings - enable/disable project's DataRefs and Commands ID usage.
 - Fixed: Bug when you try to set custom DataRef or Command key while project's DataRef or Command list is open, the key was set as ``ERROR_DATA_NOT_FOUND``
 
 2.5.0-beta (27.10.2018)